LPS-65111 by having portal-impl/src/com/liferay/portal/upgrade/v7_0_0/UpgradeModules.java extend portal-kernel/src/com/liferay/portal/kernel/upgrade/BaseUpgradeModules.java, it actually causes a breaking change. I think it's fine to just leave it in as is, and have the new classes extend portal-impl/src/com/liferay/portal/upgrade/v7_0_0/UpgradeModules.java
